requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.7
h11==0.14.0
idna==3.10
mysql-connector==2.2.9
mysql-connector-python==9.2.0
pydantic==2.10.5
pydantic_core==2.27.2
pydentic==0.0.1.dev3
python-stdnum==1.20
sniffio==1.3.1
starlette==0.45.2
typing_extensions==4.12.2
uvicorn==0.34.0
